Stalled, so committing my recent changes for posterity.
I've got the tray pretty well started, and a service to store and
retrieve launcher items. Really just need to hook up selecting and app
and adding it to the tray by either selecting and empty tray item and
then a app list item, or long clicking the app list item and then
selecting a tray item.